Move models.py into core folder
Did this because I liked the cleaner structure of the dcs/ucs repos
