+++ covid_microservice/core/models.py
